refactor(agentry-admin): migrate CLI to cobra and split into per-command files
Addresses #8. The single 1151-line main.go had grown crowded and made HTTP
calls directly from command dispatch wired up in main(), which was hard to
follow and extend.

Split the package into focused files within package main:
- main.go    tiny entrypoint (rootCmd.Execute)
- root.go    root command + persistent global flags
- client.go  HTTP request helpers (moved out of main)
- types.go   API request/response structs
- schema.go  schema command + subcommands
- agent.go   agent command + subcommands
- inbox.go   inbox command + subcommands

Replace the hand-rolled stdlib flag parsing (including the manual
global-flag scanning loop) with spf13/cobra. Auto-generated help/usage
replaces printUsage(). All commands, flags, short aliases, output strings,
and exit codes are preserved; only help text formatting changes.

Update Makefile ADMIN_MAIN_PATH to build the package directory so all
split files compile.
